Fix links to PDFs in ROCm 4.5 documentation
With the release of ROCm 5.0, the PDF manuals files for ROCm 4.5 have
been deleted from the master branch of the ROCm GitHub repository. This
deletion broke the links to the ROCm 4.5 manuals on rocmdocs.

We can fix the links by explicitly linking to the rocm-4.5.2 version
of the ROCm GitHub repository.